Rename Test component to Main and delete old Main component
