+++ src/utils/createApolloClient.js
@@ -51,7 +51,6 @@ const createApolloClient = (authStorage) => {
 
   return new ApolloClient({
     link: authLink.concat(httpLink),
-    // link: httpLink,
     cache,
   });
 };
